requirements.txt
absl-py==1.4.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
bokeh==3.2.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==1.2.0
fastjsonschema==2.16.3
ffmpeg==1.4
filelock==3.10.7
fonttools==4.39.3
fqdn==1.5.1
google-auth==2.19.1
google-auth-oauthlib==1.0.0
grpcio==1.54.2
idna==3.4
imageio==2.28.0
imbalanced-learn==0.10.1
imblearn==0.0
imutils==0.5.4
ipykernel==6.23.0
ipympl==0.9.3
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.0
jupyter-nbextensions-configurator==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.23.0
kiwisolver==1.4.4
lab==7.3
lazy_loader==0.2
livelossplot==0.5.5
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
MouseInfo==0.1.3
mpmath==1.3.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.4.13
platformdirs==3.5.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycparser==2.21
pygame==2.2.0
PyGetWindow==0.0.9
Pygments==2.15.1
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
pytesseract==0.3.10
python-dateutil==2.8.2
python-json-logger==2.0.7
pytweening==1.0.4
pytz==2022.7.1
PyWavelets==1.4.1
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.1
scikit-multilearn==0.2.0
scipy==1.10.0
seaborn==0.12.2
Send2Trash==1.8.2
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
torch==2.0.0
torchvision==0.15.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
txt2tags==3.8
typing_extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.120
uri-template==1.2.0
urllib3==1.26.15
vboxapi==1.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.6
widgetsnbextension==4.0.7
xyzservices==2023.7.0
yarg==0.1.9
